Create a table of contents alternative
A table of contents (especially if sticky) gives readers a summary of what's on the page as well as a quick way to jump around.  The Minimal Mistakes toc capability doesn't work on spash layout pages.  So instead, we use the MM `feature_row` helper capability to construct a kind of alternative toc.

If we decide we don't like it, it can easily be removed.